% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_of_the_package.R
\name{convert_to_grey}
\alias{convert_to_grey}
\title{Convert pixel to greyscale}
\usage{
convert_to_grey(rgb)
}
\arguments{
\item{rgb}{A numeric vector of length 3 representing red, green, and blue values.}
}
\value{
A raw vector of length 3 with equal gray values in R, G, and B.
}
\description{
This function takes an RGB pixel vector and returns a greyscale version.
}
